refactor: Update import paths for error handling and validator service
