+++ src/modules/auth/auth.ts
@@ -3,14 +3,14 @@ import {
   Sdk,
   SignedWalletNonceInput,
 } from '../../../gatewaySdk/sources/GatewayV3';
-import { Chain, SignCipherEnum } from '../../types/types';
-import { errorHandler } from '../../utils/error-handler';
+import { Chain, SignCipherEnum } from '../../common/enums';
+import { errorHandler } from '../../helpers/error-handler';
 import {
   isDIDValid,
   isStringValid,
   isWalletAddressValid,
   validateObjectProperties,
-} from '../../utils/validator-service';
+} from '../../common/validator-service';
 
 export class Auth {
   public sdk: Sdk;
