+++ src/modules/request/request.ts
@@ -6,11 +6,11 @@ import {
   requestsReceived_queryQueryVariables,
   requestsSent_queryQueryVariables,
 } from '../../../gatewaySdk/sources/GatewayV3';
-import { errorHandler } from '../../utils/error-handler';
+import { errorHandler } from '../../helpers/error-handler';
 import {
   isUUIDValid,
   validateObjectProperties,
-} from '../../utils/validator-service';
+} from '../../common/validator-service';
 
 export class Request {
   public sdk: Sdk;
