--- src/utils/constants.ts
@@ -1,3 +1,0 @@
-export const STRING_VALIDATION_LENGTH = 2;
-export const TEST_DEFAULT_TIMEOUT = 10000;
-export const DEFAULT_TAKE_LIMIT = 15;
